requirements.txt
torch
Pillow
opencv-python
numpy
streamlit
ultralytics
Pillow
pandas
scikit-learn
